Moved MBDataAccess class to support BB entities
